requirements.txt
aiogram==2.25.1
python-dotenv==1.0.0
supabase==2.3.1
requests==2.31.0
python-jose==3.3.0
aiohttp==3.9.1
zoom-api==0.3.4
bitrix24-python3-client==1.0.15
hashlib==20.3.0
PyJWT==2.8.0